# internal/model/timezone.go
package model

//TimeZoneUTC TimeZone = "UTC"

type TimeZone = string
const (
	UTC TimeZone = "UTC"

	UTCp1 TimeZone = "UTC+1"
	UTCp2 TimeZone = "UTC+2"
	UTCp3 TimeZone = "UTC+3"
	UTCp4 TimeZone = "UTC+4"
	UTCp5 TimeZone = "UTC+5"
	UTCp6 TimeZone = "UTC+6"
	UTCp7 TimeZone = "UTC+7"
	UTCp8 TimeZone = "UTC+8"
	UTCp9 TimeZone = "UTC+9"
	UTCp10 TimeZone = "UTC+10"
	UTCp11 TimeZone = "UTC+11"
	UTCp12 TimeZone = "UTC+12"

	UTCm1 TimeZone = "UTC-1"
	UTCm2 TimeZone = "UTC-2"
	UTCm3 TimeZone = "UTC-3"
	UTCm4 TimeZone = "UTC-4"
	UTCm5 TimeZone = "UTC-5"
	UTCm6 TimeZone = "UTC-6"
	UTCm7 TimeZone = "UTC-7"
	UTCm8 TimeZone = "UTC-8"
	UTCm9 TimeZone = "UTC-9"
	UTCm10 TimeZone = "UTC-10"
	UTCm11 TimeZone = "UTC-11"
	UTCm12 TimeZone = "UTC-12"
)